pyproject.toml
[project]
name = "voila"
version = "0.1.0"
description = "nasdaq geographic visualization"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow>=2.9.0",
    "asyncpg>=0.29.0",
    "backoff>=2.2.1",
    "colorama>=0.4.6",
    "duckdb>=1.3.2",
    "fastapi[standard]",
    "geopy>=2.4.1",
    "pandas>=2.3.1",
    "pandas-market-calendars>=4.4.2",
    "polars>=1.31.0",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=21.0.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.1",
    "redis>=6.2.0",
    "requests>=2.32.4",
    "slowapi>=0.1.9",
    "tqdm>=4.67.1",
    "yfinance>=0.2.65",
]

[dependency-groups]
dev = [
    "ruff>=0.12.7",
    "ty>=0.0.1a16",
]

[tool.ruff]
line-length = 80

[tool.uv]
package = true

[tool.setuptools]
packages = ["shared", "scrape", "app"]
